Refactor csv import
