+++ rotkehlchen/data_import/importers/cointracking.py
@@ -1,0 +1,205 @@
+import csv
+from itertools import count
+from pathlib import Path
+from typing import Any, Dict, List
+
+from rotkehlchen.assets.converters import LOCATION_TO_ASSET_MAPPING
+from rotkehlchen.assets.utils import symbol_to_asset_or_token
+from rotkehlchen.constants import ZERO
+from rotkehlchen.constants.assets import A_USD
+from rotkehlchen.data_import.utils import BaseExchangeImporter, UnsupportedCSVEntry
+from rotkehlchen.db.drivers.gevent import DBCursor
+from rotkehlchen.errors.asset import UnknownAsset
+from rotkehlchen.errors.misc import InputError
+from rotkehlchen.errors.serialization import DeserializationError
+from rotkehlchen.exchanges.data_structures import AssetMovement, Trade
+from rotkehlchen.serialization.deserialize import (
+    deserialize_asset_amount,
+    deserialize_asset_amount_force_positive,
+    deserialize_asset_movement_category,
+    deserialize_fee,
+    deserialize_timestamp_from_date,
+)
+from rotkehlchen.types import AssetAmount, AssetMovementCategory, Fee, Location, Price, TradeType
+
+
+def remap_header(fieldnames: List[str]) -> List[str]:
+    cur_count = count(1)
+    mapping = {1: 'Buy', 2: 'Sell', 3: 'Fee'}
+    return [f'Cur.{mapping[next(cur_count)]}' if f.startswith('Cur.') else f for f in fieldnames]
+
+
+def exchange_row_to_location(entry: str) -> Location:
+    """Takes the exchange row entry of Cointracking exported trades list and returns a location"""
+    if entry == 'no exchange':
+        return Location.EXTERNAL
+    if entry == 'Kraken':
+        return Location.KRAKEN
+    if entry == 'Poloniex':
+        return Location.POLONIEX
+    if entry == 'Bittrex':
+        return Location.BITTREX
+    if entry == 'Binance':
+        return Location.BINANCE
+    if entry == 'Bitmex':
+        return Location.BITMEX
+    if entry == 'Coinbase':
+        return Location.COINBASE
+    if entry in ('CoinbasePro', 'GDAX'):
+        return Location.COINBASEPRO
+    # TODO: Check if this is the correct string for Gemini from cointracking
+    if entry == 'Gemini':
+        return Location.GEMINI
+    if entry == 'Bitstamp':
+        return Location.BITSTAMP
+    if entry == 'Bitfinex':
+        return Location.BITFINEX
+    if entry == 'KuCoin':
+        return Location.KUCOIN
+    if entry == 'ETH Transaction':
+        raise UnsupportedCSVEntry(
+            'Not importing ETH Transactions from Cointracking. Cointracking does not '
+            'export enough data for them. Simply enter your ethereum accounts and all '
+            'your transactions will be auto imported directly from the chain',
+        )
+    if entry == 'BTC Transaction':
+        raise UnsupportedCSVEntry(
+            'Not importing BTC Transactions from Cointracking. Cointracking does not '
+            'export enough data for them. Simply enter your BTC accounts and all '
+            'your transactions will be auto imported directly from the chain',
+        )
+    return Location.EXTERNAL
+
+
+class CointrackingImporter(BaseExchangeImporter):
+    def _consume_cointracking_entry(
+            self,
+            cursor: DBCursor,
+            csv_row: Dict[str, Any],
+            timestamp_format: str = '%d.%m.%Y %H:%M:%S',
+    ) -> None:
+        """Consumes a cointracking entry row from the CSV and adds it into the database
+        Can raise:
+            - DeserializationError if something is wrong with the format of the expected values
+            - UnsupportedCSVEntry if importing of this entry is not supported.
+            - IndexError if the CSV file is corrupt
+            - KeyError if the an expected CSV key is missing
+            - UnknownAsset if one of the assets founds in the entry are not supported
+        """
+        row_type = csv_row['Type']
+        timestamp = deserialize_timestamp_from_date(
+            date=csv_row['Date'],
+            formatstr=timestamp_format,
+            location='cointracking.info',
+        )
+        location = exchange_row_to_location(csv_row['Exchange'])
+        asset_resolver = LOCATION_TO_ASSET_MAPPING.get(location, symbol_to_asset_or_token)
+        notes = csv_row['Comment']
+        if location == Location.EXTERNAL:
+            notes += f'. Data from -{csv_row["Exchange"]}- not known by rotki.'
+
+        fee = Fee(ZERO)
+        fee_currency = A_USD  # whatever (used only if there is no fee)
+        if csv_row['Fee'] != '':
+            fee = deserialize_fee(csv_row['Fee'])
+            fee_currency = asset_resolver(csv_row['Cur.Fee'])
+
+        if row_type in ('Gift/Tip', 'Trade', 'Income'):
+            base_asset = asset_resolver(csv_row['Cur.Buy'])
+            quote_asset = None if csv_row['Cur.Sell'] == '' else asset_resolver(csv_row['Cur.Sell'])  # noqa: E501
+            if quote_asset is None and row_type not in ('Gift/Tip', 'Income'):
+                raise DeserializationError('Got a trade entry with an empty quote asset')
+
+            if quote_asset is None:
+                # Really makes no difference as this is just a gift and the amount is zero
+                quote_asset = A_USD
+            base_amount_bought = deserialize_asset_amount(csv_row['Buy'])
+            if base_amount_bought == ZERO:
+                raise DeserializationError('Bought amount in trade is zero')
+
+            if csv_row['Sell'] != '-':
+                quote_amount_sold = deserialize_asset_amount(csv_row['Sell'])
+            else:
+                quote_amount_sold = AssetAmount(ZERO)
+            rate = Price(quote_amount_sold / base_amount_bought)
+
+            trade = Trade(
+                timestamp=timestamp,
+                location=location,
+                base_asset=base_asset,
+                quote_asset=quote_asset,
+                trade_type=TradeType.BUY,  # It's always a buy during cointracking import
+                amount=base_amount_bought,
+                rate=rate,
+                fee=fee,
+                fee_currency=fee_currency,
+                link='',
+                notes=notes,
+            )
+            self.add_trade(cursor, trade)
+        elif row_type in ('Deposit', 'Withdrawal'):
+            category = deserialize_asset_movement_category(row_type.lower())
+            if category == AssetMovementCategory.DEPOSIT:
+                amount = deserialize_asset_amount(csv_row['Buy'])
+                asset = asset_resolver(csv_row['Cur.Buy'])
+            else:
+                amount = deserialize_asset_amount_force_positive(csv_row['Sell'])
+                asset = asset_resolver(csv_row['Cur.Sell'])
+
+            asset_movement = AssetMovement(
+                location=location,
+                category=category,
+                address=None,
+                transaction_id=None,
+                timestamp=timestamp,
+                asset=asset,
+                amount=amount,
+                fee=fee,
+                fee_asset=fee_currency,
+                link='',
+            )
+            self.add_asset_movement(cursor, asset_movement)
+        else:
+            raise UnsupportedCSVEntry(
+                f'Unknown entrype type "{row_type}" encountered during cointracking '
+                f'data import. Ignoring entry',
+            )
+
+    def _import_csv(
+            self,
+            cursor: DBCursor,
+            filepath: Path,
+            **kwargs: Any,
+    ) -> None:
+        """May raise:
+        - InputError if one of the rows is malformed
+        """
+        with open(filepath, 'r', encoding='utf-8-sig') as csvfile:
+            data = csv.reader(csvfile, delimiter=',', quotechar='"')
+            header = remap_header(next(data))
+            for row in data:
+                try:
+                    self._consume_cointracking_entry(cursor, dict(zip(header, row)), **kwargs)
+                except UnknownAsset as e:
+                    self.db.msg_aggregator.add_warning(
+                        f'During cointracking CSV import found action with unknown '
+                        f'asset {e.asset_name}. Ignoring entry',
+                    )
+                    continue
+                except IndexError:
+                    self.db.msg_aggregator.add_warning(
+                        'During cointracking CSV import found entry with '
+                        'unexpected number of columns',
+                    )
+                    continue
+                except DeserializationError as e:
+                    self.db.msg_aggregator.add_warning(
+                        f'Error during cointracking CSV import deserialization. '
+                        f'Error was {str(e)}. Ignoring entry',
+                    )
+                    continue
+                except UnsupportedCSVEntry as e:
+                    self.db.msg_aggregator.add_warning(str(e))
+                    continue
+                except KeyError as e:
+                    raise InputError(f'Could not find key {str(e)} in csv row {str(row)}') from e
