+++ rotkehlchen/rotkehlchen.py
@@ -52,7 +52,7 @@ from rotkehlchen.chain.substrate.utils import (
 from rotkehlchen.config import default_data_directory
 from rotkehlchen.constants.misc import ONE, ZERO
 from rotkehlchen.data_handler import DataHandler
-from rotkehlchen.data_import.manager import DataImporter
+from rotkehlchen.data_import.manager import CSVDataImporter
 from rotkehlchen.data_migrations.manager import DataMigrationManager
 from rotkehlchen.db.settings import DBSettings, ModifiableDBSettings
 from rotkehlchen.errors.api import PremiumAuthenticationError
@@ -203,7 +203,7 @@ class Rotkehlchen():
             exception_is_error=False,
             method=self.data.db.ensure_data_integrity,
         )
-        self.data_importer = DataImporter(db=self.data.db)
+        self.data_importer = CSVDataImporter(db=self.data.db)
         self.premium_sync_manager = PremiumSyncManager(data=self.data, password=password)
         # set the DB in the external services instances that need it
         self.cryptocompare.set_database(self.data.db)
